Add a keep-alive NOOP to the C-Gate connection
At times, the C-Gate connection may be interrupted due to connectivity issues (i.e. WiFi dropouts) or because the server crashed or was restarted. We won't find out about this until we attempt to write to the socket possibly hours after the connection was dropped.

Add a periodic NOOP command on the C-Gate connection to detect a failed connection sooner.